+++ lib/cyclonedx/cocoapods/pod.rb
@@ -37,13 +37,6 @@ module CycloneDX
         self
       end
 
-      def purl
-        purl_name = CGI.escape(name.split('/').first)
-        source_qualifier = source.nil? || source.source_qualifier.empty? ? '' : "?#{source.source_qualifier.map { |key, value| "#{key}=#{CGI.escape(value)}" }.join('&')}"
-        purl_subpath = name.split('/').length > 1 ? "##{name.split('/').drop(1).map { |component| CGI.escape(component) }.join('/')}" : ''
-        return "pkg:cocoapods/#{purl_name}@#{CGI.escape(version.to_s)}#{source_qualifier}#{purl_subpath}"
-      end
-
       def to_s
         "Pod<#{name}, #{version.to_s}>"
       end
