+++ lib/cyclonedx/cocoapods/source.rb
@@ -2,17 +2,10 @@ module CycloneDX
   module CocoaPods
     module Source
       class CocoaPodsRepository
-        LEGACY_REPOSITORY = 'https://github.com/CocoaPods/Specs.git'.freeze
-        CDN_REPOSITORY = 'trunk'.freeze
-
         attr_reader :url
 
         def initialize(url:)
           @url = url
-        end
-
-        def source_qualifier
-          url == LEGACY_REPOSITORY || url == CDN_REPOSITORY ? {} : { repository_url: url }
         end
       end
 
@@ -25,20 +18,11 @@ module CycloneDX
           raise ArgumentError, "Invalid checkout information" if !type.nil? && !VALID_TYPES.include?(type)
           @url, @type, @label = url, type, label
         end
-
-        def source_qualifier
-          { vcs_url: (label.nil? ? url : "#{url}@#{label}") }
-        end
       end
 
       class LocalPod
         def initialize(path:)
           @path = path
-        end
-
-        def source_qualifier
-          # TODO: Should we generate a source qualifier for :path dependencies?
-          {}
         end
       end
 
@@ -47,10 +31,6 @@ module CycloneDX
 
         def initialize(url:)
           @url = url
-        end
-
-        def source_qualifier
-          { download_url: @url }
         end
       end
     end
